pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.12",
    "fastapi>=0.115.9",
    "langchain-openai>=0.3.23",
    "langchain>=0.3.25",
    "openai>=1.84.0",
    "pydantic>=2.11.5",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
    "langchain-community>=0.3.25",
    "langchain-text-splitters>=0.3.8",
    "requests>=2.32.3",
    "pypdf2>=3.0.1",
    "langsmith>=0.3.45",
]
